default to float image output when transfoming images. Avoid unsigned signed conversion problems.

+++ src/main/java/ch/epfl/biop/wrappers/elastix/RegParamAffine_Default.java
@@ -39,7 +39,7 @@ public class RegParamAffine_Default extends RegistrationParameters {
 		FinalBSplineInterpolationOrder=3;
 		DefaultPixelValue=0f;
 		WriteResultImage=false;
-		ResultImagePixelType="short";
+		ResultImagePixelType="float";
 		ResultImageFormat="tif";
 	}
 }
